pyproject.toml
[project]
name = "fast-video-object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi==2022.12.7",
    "charset-normalizer==2.1.1",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "filelock==3.13.1",
    "fonttools==4.56.0",
    "fsspec==2024.6.1",
    "idna==3.4",
    "jinja2==3.1.4",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "mpmath==1.3.0",
    "networkx==3.3",
    "numpy==2.1.2",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "opencv-python==4.11.0.86",
    "packaging==24.1",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "pip==25.0.1",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "requests==2.28.1",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "setuptools==70.2.0",
    "six==1.17.0",
    "sympy==1.13.1",
    "torch==2.6.0+cu126",
    "torchaudio==2.6.0+cu126",
    "torchcodec==0.2.0+cu126",
    "torchvision==0.21.0+cu126",
    "tqdm==4.66.5",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "ultralytics==8.3.23",
    "ultralytics-thop==2.0.14",
    "urllib3==1.26.13",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
torchcodec = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
